fix(api): split build-payload for file-length gate (#301)
Extract graph-payload-loaders.ts and graph-payload-types.ts so all
graph assembly modules stay under the 300-line cap.
